[rhi] Starting on conversions to the new pipeline create info struct
