+++ src/nova_renderer.cpp
@@ -102,6 +102,66 @@ void rex_fini() {
 }
 
 namespace nova::renderer {
+    struct RX_HINT_EMPTY_BASES BackbufferOutputPipelineCreateInfo : PipelineStateCreateInfo {
+        BackbufferOutputPipelineCreateInfo();
+    };
+
+    BackbufferOutputPipelineCreateInfo::BackbufferOutputPipelineCreateInfo() {
+        name = BACKBUFFER_OUTPUT_PIPELINE_NAME;
+
+        const rx::string vertex_source{R"(
+            struct VsInput {
+                float2 position : POSITION;
+            };
+
+            struct VsOutput {
+                float4 position : SV_POSITION;
+                float2 uv : TEXCOORD;
+
+            };
+
+            VsOutput main(VsInput input) {
+                VsOutput output;
+                output.position = float4(input.position, 0, 1);
+                output.uv = input.position; // TODO: Figure out if I should flip the texture coordinates
+
+                return output;
+            })"};
+        const auto& vertex_spirv = shaderpack::compile_shader(vertex_source, rhi::ShaderStage::Vertex, rhi::ShaderLanguage::Hlsl);
+        vertex_shader = {"/nova/shaders/backbuffer_output.vertex.hlsl", vertex_spirv};
+
+        const rx::string pixel_source{R"(
+            [[vk::binding(0, 0)]]
+            Texture2D ui_output : register(t0);
+
+            [[vk::binding(1, 0)]]
+            Texture2D scene_output : register(t1);
+
+            [[vk::binding(2, 0)]]
+            SamplerState sampler : register(s0);
+
+            struct VsOutput {
+                float4 position : SV_POSITION;
+                float2 us : TEXCOORD;
+            };
+
+            float3 main(VsOutput input) : SV_Target {
+                float4 ui_color = ui_output.Sample(sampler, input.uv);
+                float4 scene_color = scene_output.Sample(sampler, input.uv);
+
+                float3 combined_color = lerp(ui_color.rgb, scene_color.rgb, ui_color.a);
+
+                return combined_color;
+            })"};
+        const auto& pixel_spirv = shaderpack::compile_shader(pixel_source, rhi::ShaderStage::Fragment, rhi::ShaderLanguage::Hlsl);
+        pixel_shader = {"/nova/shaders/backbuffer_output.pixel.hlsl", pixel_spirv};
+
+        // TODO: Figure out how to make the input textures into input attachments
+        color_attachments.emplace_back(BACKBUFFER_NAME, shaderpack::PixelFormatEnum::RGBA8, false);
+    }
+
+    RX_GLOBAL<BackbufferOutputPipelineCreateInfo> backbuffer_output_pipeline_create_info{"Nova", "BackbufferOutputPipelineCreateInfo"};
+
     bool FullMaterialPassName::operator==(const FullMaterialPassName& other) const {
         return material_name == other.material_name && pass_name == other.pass_name;
     }
@@ -432,8 +492,10 @@ namespace nova::renderer {
         }
 
         pipeline_create_infos.each_fwd([&](const shaderpack::PipelineData& pipeline_create_info) {
-            if(pipeline_storage->create_pipeline(pipeline_create_info)) {
-                auto pipeline = pipeline_storage->get_pipeline(pipeline_create_info.name);
+            const auto pipeline_state_create_info = to_pipeline_state_create_info(pipeline_create_info);
+
+            if(pipeline_storage->create_pipeline(pipeline_state_create_info)) {
+                auto pipeline = pipeline_storage->get_pipeline(pipeline_state_create_info.name);
                 create_materials_for_pipeline(*pipeline, materials, pipeline_create_info.name);
             }
         });
@@ -787,7 +849,7 @@ namespace nova::renderer {
 
     void NovaRenderer::create_renderpass_manager() { rendergraph = global_allocator->create<Rendergraph>(global_allocator, *device); }
 
-    void NovaRenderer::create_builtin_renderpasses()  {
+    void NovaRenderer::create_builtin_renderpasses() {
         if(rendergraph->create_renderpass<BackbufferOutputRenderpass>(*device_resources, this) == nullptr) {
             logger(rx::log::level::k_error, "Could not create the backbuffer output renderpass");
         }
